Move test specs back to correct file
